Remove TfOpRoofLineCostEstimator
This is the last remaining dependency on Tensorflow, and removal will allow us to drastically simplify our builds and dependencies.
This results in no longer estimating TFOp peak HBM and FLOPs. It is my understanding that this feature has little visible effect, and is not being used by anyone. If this change has impacted your workflow, please open an issue on the repo and we'll investigate alternative ways to obtain this information.

PiperOrigin-RevId: 751205020