Add source stack trace source code view to Trace Viewer page.
We use the new stack trace page component to show snippets of source code around the stack trace in a new page. This is particularly useful for Trace Viewer, since it is not written in Angular, but `StackTraceSnippet` is written in Angular. This component provides a bridge between the two.

The implementation is quite different from all the previous changes that add the source code viewer to other tools. This is because Trace Viewer is not an Angular component and a lot needed to be done to bride the gap.

PiperOrigin-RevId: 780211944