Remove duplicate version checks
These checks have caused a number of issues so far, and the main reason that they were added is to improve the clarity of error logging from tensorboard. At this point, we can rely on tensorboard to error out for us, as this check has continually been inconsistent.

Addresses https://github.com/openxla/xprof/issues/1714

PiperOrigin-RevId: 796507917